+++ CHANGELOG.md
@@ -47,6 +47,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - **CUMULUS-3518**
   - Update existing usage of `@cumulus/aws-client` lambda service to use AWS SDK v3 `send` syntax
+- **CUMULUS-2892**
+  - Updates `aws-client`'s EC2 client to use AWS SDK v3.
+- **CUMULUS-2896**
+  - Updated Secrets Manager code to AWS SDK v3.
 - **CUMULUS-2901**
   - Updated STS code to AWS SDK v3.
 - **CUMULUS-2898**
@@ -71,8 +75,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Minor refactor of `@cumulus/lzards-api-client` to:
     - Use proper ECMAScript import for `@cumulus/launchpad-auth`
     - Update incorrect docstring
-- **CUMULUS-2896**
-  - Updated Secrets Manager code to AWS SDK v3.
 - **CUMULUS-3449**
   - Updated `@cumulus/db` package and configure knex hook postProcessResponse to convert the return string
     from columns ending with "cumulus_id" to number.
@@ -80,6 +82,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Updated `example/cumulus-tf/orca.tf` to use v9.0.4
 - **CUMULUS-3527**
   - Added suppport for additional kex algorithms in the sftp-client.
+- **CUMULUS-3610**
+  - Updated `aws-client`'s ES client to use AWS SDK v3.
 
 ### Fixed